// inventory-backend/functions/src/Models/Product.ts
export class Product  {
    Id: any;
    PartName: any;
    CategoryId: any;
    SubCategoryId: any;
    ManufacturerPartNumber: any;
    ManufacturerName: any;
    DatasheetLink: any;
    Size: any;
    EIACaseCode: any;
    VoltageRatings: any;
    RatingTolarence: any;
    TemperatureCoefficientId: any;
    SupplierId: any;
    Remarks: any;
}